+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## latest
 
-- ...
+- changed to RxJS 5.0.0 beta1 from RxJS4
 
 ## 0.0.4